+++ ChestsAnywhere/Framework/ContainerData.cs
@@ -1,5 +1,11 @@
+using System;
+using System.Diagnostics.CodeAnalysis;
+using System.Globalization;
 using System.Text.RegularExpressions;
 using Pathoschild.Stardew.Automate.Framework;
+using Pathoschild.Stardew.ChestsAnywhere.Framework.Containers;
+using Pathoschild.Stardew.Common;
+using StardewValley;
 
 namespace Pathoschild.Stardew.ChestsAnywhere.Framework
 {
@@ -11,6 +17,9 @@ namespace Pathoschild.Stardew.ChestsAnywhere.Framework
         *********/
         /// <summary>A regular expression which matches a group of tags in the chest name.</summary>
         private const string TagGroupPattern = @"\|([^\|]+)\|";
+
+        /// <summary>The key prefix with which to store container options in a <see cref="ModDataDictionary"/>.</summary>
+        private const string ModDataPrefix = "Pathoschild.ChestsAnywhere";
 
 
         /*********
@@ -42,6 +51,7 @@ namespace Pathoschild.Stardew.ChestsAnywhere.Framework
         ** Public methods
         *********/
         /// <summary>Construct an empty instance.</summary>
+        [SuppressMessage("ReSharper", "UnusedMember.Global", Justification = "Used in deserialization for " + nameof(ShippingBinContainer))]
         public ContainerData() { }
 
         /// <summary>Construct an empty instance.</summary>
@@ -51,10 +61,86 @@ namespace Pathoschild.Stardew.ChestsAnywhere.Framework
             this.DefaultInternalName = defaultInternalName;
         }
 
+        /// <summary>Load contain data for the given item.</summary>
+        /// <param name="item">The item whose container data to load.</param>
+        /// <param name="defaultDisplayName">The default display name.</param>
+        public static ContainerData GetFor(Item item, string defaultDisplayName)
+        {
+            ModDataDictionary data = item.modData;
+            return new ContainerData(defaultDisplayName)
+            {
+                IsIgnored = data.ReadField($"{ContainerData.ModDataPrefix}/{nameof(ContainerData.IsIgnored)}", bool.Parse),
+                Category = data.ReadField($"{ContainerData.ModDataPrefix}/{nameof(ContainerData.Category)}"),
+                Name = data.ReadField($"{ContainerData.ModDataPrefix}/{nameof(ContainerData.Name)}"),
+                Order = data.ReadField($"{ContainerData.ModDataPrefix}/{nameof(ContainerData.Order)}", int.Parse),
+                AutomateStoreItems = data.ReadField(AutomateContainerHelper.StoreItemsKey, p => (AutomateContainerPreference)Enum.Parse(typeof(AutomateContainerPreference), p), defaultValue: AutomateContainerPreference.Allow),
+                AutomateTakeItems = data.ReadField(AutomateContainerHelper.TakeItemsKey, p => (AutomateContainerPreference)Enum.Parse(typeof(AutomateContainerPreference), p), defaultValue: AutomateContainerPreference.Allow)
+            };
+        }
+
+        /// <summary>Save the container data to the given mod data.</summary>
+        /// <param name="data">The mod data.</param>
+        public void ToModData(ModDataDictionary data)
+        {
+            data
+                .WriteField($"{ContainerData.ModDataPrefix}/{nameof(ContainerData.IsIgnored)}", this.IsIgnored ? "true" : null)
+                .WriteField($"{ContainerData.ModDataPrefix}/{nameof(ContainerData.Category)}", this.Category)
+                .WriteField($"{ContainerData.ModDataPrefix}/{nameof(ContainerData.Name)}", !this.HasDefaultDisplayName() ? this.Name : null)
+                .WriteField($"{ContainerData.ModDataPrefix}/{nameof(ContainerData.Order)}", this.Order != 0 ? this.Order?.ToString(CultureInfo.InvariantCulture) : null)
+                .WriteField(AutomateContainerHelper.StoreItemsKey, this.AutomateStoreItems != AutomateContainerPreference.Allow ? this.AutomateStoreItems.ToString() : null)
+                .WriteField(AutomateContainerHelper.TakeItemsKey, this.AutomateTakeItems != AutomateContainerPreference.Allow ? this.AutomateTakeItems.ToString() : null);
+        }
+
+        /// <summary>Whether the container has the default display name.</summary>
+        public bool HasDefaultDisplayName()
+        {
+            return string.IsNullOrWhiteSpace(this.Name) || this.Name == this.DefaultInternalName;
+        }
+
+        /// <summary>Get whether the container has any non-default data.</summary>
+        public bool HasData()
+        {
+            return
+                !this.HasDefaultDisplayName()
+                || (this.Order.HasValue && this.Order != 0)
+                || this.IsIgnored
+                || !string.IsNullOrWhiteSpace(this.Category)
+                || this.AutomateTakeItems != AutomateContainerPreference.Allow
+                || this.AutomateStoreItems != AutomateContainerPreference.Allow;
+        }
+
+        /// <summary>Reset all container data to the default.</summary>
+        public void Reset()
+        {
+            this.Name = this.DefaultInternalName;
+            this.Order = null;
+            this.IsIgnored = false;
+            this.Category = null;
+            this.AutomateTakeItems = AutomateContainerPreference.Allow;
+            this.AutomateStoreItems = AutomateContainerPreference.Allow;
+        }
+
+        /// <summary>Load contain data for the given item, migrating it if needed.</summary>
+        /// <param name="item">The item whose container data to load.</param>
+        /// <param name="defaultDisplayName">The default display name.</param>
+        public static void MigrateLegacyData(Item item, string defaultDisplayName)
+        {
+            if (string.IsNullOrWhiteSpace(item.Name) || item.Name == defaultDisplayName)
+                return;
+
+            ContainerData data = ContainerData.ParseLegacyDataFromName(item.Name, defaultDisplayName);
+            item.Name = defaultDisplayName;
+            data.ToModData(item.modData);
+        }
+
+
+        /*********
+        ** Private methods
+        *********/
         /// <summary>Parse a serialized name string.</summary>
         /// <param name="name">The serialized name string.</param>
         /// <param name="defaultDisplayName">The default display name for the container.</param>
-        public static ContainerData ParseName(string name, string defaultDisplayName)
+        private static ContainerData ParseLegacyDataFromName(string name, string defaultDisplayName)
         {
             // construct instance
             ContainerData data = new ContainerData(defaultDisplayName);
@@ -96,68 +182,6 @@ namespace Pathoschild.Stardew.ChestsAnywhere.Framework
                 : defaultDisplayName;
 
             return data;
-        }
-
-        /// <summary>Get a serialized name representation of the container data.</summary>
-        public string ToName()
-        {
-            // name
-            string internalName = !this.HasDefaultDisplayName() ? this.Name : this.DefaultInternalName;
-
-            // order
-            if (this.Order.HasValue && this.Order != 0)
-                internalName += $" |{this.Order}|";
-
-            // ignore
-            if (this.IsIgnored)
-                internalName += " |ignore|";
-
-            // category
-            if (!string.IsNullOrWhiteSpace(this.Category))
-                internalName += $" |cat:{this.Category}|";
-
-            // Automate input
-            if (!this.AutomateStoreItems.IsAllowed())
-                internalName += " |automate:no-store|";
-            else if (this.AutomateStoreItems.IsPreferred())
-                internalName += " |automate:prefer-store|";
-
-            // Automate output
-            if (!this.AutomateTakeItems.IsAllowed())
-                internalName += " |automate:no-take|";
-            else if (this.AutomateTakeItems.IsPreferred())
-                internalName += " |automate:prefer-take|";
-
-            return internalName;
-        }
-
-        /// <summary>Whether the container has the default display name.</summary>
-        public bool HasDefaultDisplayName()
-        {
-            return string.IsNullOrWhiteSpace(this.Name) || this.Name == this.DefaultInternalName;
-        }
-
-        /// <summary>Get whether the container has any non-default data.</summary>
-        public bool HasData()
-        {
-            return
-                !this.HasDefaultDisplayName()
-                || (this.Order.HasValue && this.Order != 0)
-                || this.IsIgnored
-                || !string.IsNullOrWhiteSpace(this.Category)
-                || this.AutomateTakeItems != AutomateContainerPreference.Allow
-                || this.AutomateStoreItems != AutomateContainerPreference.Allow;
-        }
-
-        /// <summary>Reset all container data to the default.</summary>
-        public void Reset()
-        {
-            this.Name = this.DefaultInternalName;
-            this.Order = null;
-            this.IsIgnored = false;
-            this.Category = null;
-            this.AutomateTakeItems = AutomateContainerPreference.Allow;
-            this.AutomateStoreItems = AutomateContainerPreference.Allow;
         }
     }
 }