+++ ChestsAnywhere/ModEntry.cs
@@ -57,6 +57,7 @@ namespace Pathoschild.Stardew.ChestsAnywhere
 
             // hook events
             helper.Events.GameLoop.SaveLoaded += this.OnSaveLoaded;
+            helper.Events.World.LocationListChanged += this.OnLocationListChanged;
             helper.Events.GameLoop.UpdateTicked += this.OnUpdateTicked;
             helper.Events.GameLoop.UpdateTicking += this.OnUpdateTicking;
             helper.Events.Display.RenderedHud += this.OnRenderedHud;
@@ -87,6 +88,22 @@ namespace Pathoschild.Stardew.ChestsAnywhere
             // show multiplayer limitations warning
             if (!Context.IsMainPlayer)
                 this.Monitor.Log("Multiplayer limitations: you can only access chests in synced locations since you're not the main player. This is due to limitations in the game's sync logic.", LogLevel.Info);
+
+            // migrate legacy chest data
+            if (Context.IsMainPlayer)
+                this.MigrateLegacyData();
+        }
+
+        /// <summary>The method invoked after locations are added or removed from the game.</summary>
+        /// <param name="sender">The event sender.</param>
+        /// <param name="e">The event arguments.</param>
+        private void OnLocationListChanged(object sender, LocationListChangedEventArgs e)
+        {
+            if (Context.IsMultiplayer)
+            {
+                foreach (GameLocation location in e.Added)
+                    this.MigrateLegacyData(location);
+            }
         }
 
         /// <summary>The method invoked when the interface has finished rendering.</summary>
@@ -308,6 +325,21 @@ namespace Pathoschild.Stardew.ChestsAnywhere
                 return I18n.Errors_NoChestsInRange();
 
             return I18n.Errors_NoChests();
+        }
+
+        /// <summary>Migrate legacy container data, if needed.</summary>
+        private void MigrateLegacyData()
+        {
+            foreach (var chest in this.ChestFactory.GetChests(RangeHandler.Unlimited()))
+                chest.Container.MigrateLegacyData();
+        }
+
+        /// <summary>Migrate legacy container data, if needed.</summary>
+        /// <param name="location">The location whose chests to migrate.</param>
+        private void MigrateLegacyData(GameLocation location)
+        {
+            foreach (var chest in this.ChestFactory.GetChests(RangeHandler.SpecificLocation(location)))
+                chest.Container.MigrateLegacyData();
         }
     }
 }