+++ Automate/release-notes.md
@@ -20,6 +20,13 @@ Released 21 November 2020.
 * Fixed constructed shipping bins triggering the main shipping bin's animation.
 * Fixed shipped items not added to existing stacks if possible.
 
+**Breaking changes for mod authors:**
+* If you added a `chest.Capacity` property for Automate, you should patch or override the new
+  `chest.GetActualCapacity()` method in 1.5 instead.
+* Automate previously stored automation options in the chest name using tags like
+  `|automate:no-store|`. These are now stored in the new `modData` field added in 1.5; existing
+  chests will be migrated automatically.
+
 ## 1.18.2
 Released 04 November 2020.